revert abstracted table move and subscription transformation refactor
